Search a subset of reference sequences

+++ include/stellar/options/index_options.hpp
@@ -8,7 +8,7 @@ namespace stellar
 
 struct IndexOptions
 {
-    unsigned qGram{std::numeric_limits<unsigned>::max()}; // length of the q-grams
+    size_t qGram{std::numeric_limits<size_t>::max()}; // length of the q-grams
     double qgramAbundanceCut{1};
 };
